Add a note about avoiding function hoisting #75
After discussions in dev meeting we decided to avoid hoisting to avoid confusion and PR back-and-forth. We tried adding a lint rule but it was overly aggressive so I'm adding a note here instead.

Note: I removed/replaced some old content about the difference between function declarations and function expressions since it seems less relevant if we're avoiding hoisting. I'm happy to put it back in if others feel differently.